WIP restructure to match beacon chain DB
