+++ slasher/db/BUILD.bazel
@@ -3,54 +3,20 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "attester_slashings.go",
-        "block_header.go",
+        "alias.go",
         "db.go",
-        "indexed_attestations.go",
-        "min_max_span.go",
-        "proposer_slashings.go",
-        "schema.go",
-        "setup_db.go",
-        "validator_id_pubkey.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/slasher/db",
     visibility = ["//slasher:__subpackages__"],
     deps = [
-        "//proto/slashing:go_default_library",
-        "//shared/bytesutil:go_default_library",
-        "//shared/hashutil:go_default_library",
-        "//shared/params:go_default_library",
         "//slasher/db/iface:go_default_library",
-        "//slasher/db/types:go_default_library",
-        "//slasher/flags:go_default_library",
-        "@com_github_boltdb_bolt//:go_default_library",
-        "@com_github_dgraph_io_ristretto//:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_urfave_cli//:go_default_library",
+        "//slasher/db/kv:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "attester_slashings_test.go",
-        "block_header_test.go",
-        "indexed_attestations_test.go",
-        "min_max_span_test.go",
-        "proposer_slashings_test.go",
-        "setup_db_test.go",
-        "validator_id_pubkey_test.go",
-    ],
+    srcs = ["db_test.go"],
     embed = [":go_default_library"],
-    deps = [
-        "//proto/slashing:go_default_library",
-        "//slasher/db/types:go_default_library",
-        "//slasher/flags:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_urfave_cli//:go_default_library",
-    ],
+    deps = ["//slasher/db/kv:go_default_library"],
 )
