+++ slasher/db/db_test.go
@@ -1,0 +1,5 @@
+package db
+
+import "github.com/prysmaticlabs/prysm/slasher/db/kv"
+
+var _ = Database(&kv.Store{})
