+++ slasher/db/kv/BUILD.bazel
@@ -3,18 +3,25 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "detect_update_min_max_span.go",
-        "server.go",
+        "attester_slashings.go",
+        "block_header.go",
+        "indexed_attestations.go",
+        "kv.go",
+        "min_max_span.go",
+        "proposer_slashings.go",
+        "schema.go",
+        "validator_id_pubkey.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/slasher/rpc",
-    visibility = ["//visibility:public"],
+    importpath = "github.com/prysmaticlabs/prysm/slasher/db/kv",
+    visibility = ["//slasher:__subpackages__"],
     deps = [
-        "//beacon-chain/core/helpers:go_default_library",
         "//proto/slashing:go_default_library",
+        "//shared/bytesutil:go_default_library",
         "//shared/hashutil:go_default_library",
         "//shared/params:go_default_library",
-        "//slasher/db/kv:go_default_library",
         "//slasher/db/types:go_default_library",
+        "@com_github_boltdb_bolt//:go_default_library",
+        "@com_github_dgraph_io_ristretto//:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
@@ -25,43 +32,21 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "detect_update_min_max_span_test.go",
-        "server_test.go",
-        "slashing_bench_test.go",
+        "attester_slashings_test.go",
+        "block_header_test.go",
+        "indexed_attestations_test.go",
+        "min_max_span_test.go",
+        "proposer_slashings_test.go",
+        "validator_id_pubkey_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//proto/slashing:go_default_library",
-        "//shared/params:go_default_library",
-        "//slasher/db:go_default_library",
+        "//slasher/db/testing:go_default_library",
         "//slasher/db/types:go_default_library",
         "//slasher/flags:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-        "@com_github_urfave_cli//:go_default_library",
-    ],
-)
-
-# gazelle:exclude detect_update_min_max_span_bench_test.go
-go_test(
-    name = "go_benchmark_test",
-    size = "medium",
-    srcs = ["slashing_bench_test.go"],
-    args = [
-        "-test.bench=.",
-        "-test.benchmem",
-        "-test.v",
-    ],
-    embed = [":go_default_library"],
-    local = True,
-    tags = [
-        "benchmark",
-        "manual",
-        "no-cache",
-    ],
-    deps = [
-        "//slasher/db:go_default_library",
-        "//slasher/flags:go_default_library",
         "@com_github_urfave_cli//:go_default_library",
     ],
 )
