+++ slasher/db/kv/block_header.go
@@ -1,4 +1,4 @@
-package db
+package kv
 
 import (
 	"bytes"
