+++ slasher/db/kv/indexed_attestations.go
@@ -1,4 +1,4 @@
-package db
+package kv
 
 import (
 	"bytes"
