+++ slasher/db/kv/indexed_attestations_test.go
@@ -1,4 +1,4 @@
-package db
+package kv
 
 import (
 	"flag"
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
+	testDB "github.com/prysmaticlabs/prysm/slasher/db/testing"
 	"github.com/urfave/cli"
 )
 
@@ -42,8 +43,8 @@ func TestNilDBHistoryIdxAtt(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	epoch := uint64(1)
 	validatorID := uint64(1)
@@ -69,8 +70,8 @@ func TestSaveIdxAtt(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	for _, tt := range tests {
 		err := db.SaveIndexedAttestation(tt.idxAtt)
@@ -94,8 +95,8 @@ func TestDeleteHistoryIdxAtt(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	for _, tt := range tests {
 
@@ -142,8 +143,8 @@ func TestHasIndexedAttestation(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	for _, tt := range tests {
 		exists, err := db.HasIndexedAttestation(tt.idxAtt.Data.Target.Epoch, tt.idxAtt.AttestingIndices[0])
@@ -174,8 +175,8 @@ func TestPruneHistoryIdxAtt(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	for _, tt := range tests {
 		err := db.SaveIndexedAttestation(tt.idxAtt)
