+++ slasher/db/kv/proposer_slashings.go
@@ -1,4 +1,4 @@
-package db
+package kv
 
 import (
 	"bytes"
