+++ slasher/db/kv/validator_id_pubkey.go
@@ -1,4 +1,4 @@
-package db
+package kv
 
 import (
 	"github.com/boltdb/bolt"
