+++ slasher/db/kv/validator_id_pubkey_test.go
@@ -1,10 +1,11 @@
-package db
+package kv
 
 import (
 	"bytes"
 	"flag"
 	"testing"
 
+	testDB "github.com/prysmaticlabs/prysm/slasher/db/testing"
 	"github.com/urfave/cli"
 )
 
@@ -36,8 +37,8 @@ func TestNilDBValidatorPublicKey(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	validatorID := uint64(1)
 
@@ -55,8 +56,8 @@ func TestSavePubKey(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	for _, tt := range pkTests {
 		err := db.SavePubKey(tt.validatorID, tt.pk)
@@ -80,8 +81,8 @@ func TestDeletePublicKey(t *testing.T) {
 	app := cli.NewApp()
 	set := flag.NewFlagSet("test", 0)
 	ctx := cli.NewContext(app, set, nil)
-	db := SetupSlasherDB(t, ctx)
-	defer TeardownSlasherDB(t, db)
+	db := testDB.SetupSlasherDB(t, ctx)
+	defer testDB.TeardownSlasherDB(t, db)
 
 	for _, tt := range pkTests {
