+++ slasher/db/testing/setup_db_test.go
@@ -1,4 +1,4 @@
-package db
+package testing
 
 import (
 	"flag"
